refactored for long-term scalability
